+++ fs/vfs.h
@@ -4,6 +4,7 @@
 #ifndef VFS_H
 #define VFS_H
 
+#include "types.h"
 #include <stdint.h>
 #include <stdbool.h>
 #include <stddef.h>
@@ -47,8 +48,8 @@ typedef struct vfs_fs_ops {
     // File operations — return opaque FS handle
     void* (*open)(void *fs_private, const char *rel_path, const char *mode);
     void  (*close)(void *fs_private, void *file_handle);
-    int   (*read)(void *fs_private, void *file_handle, void *buf, int size);
-    int   (*write)(void *fs_private, void *file_handle, const void *buf, int size);
+    ssize_t   (*read)(void *fs_private, void *file_handle, void *buf, size_t size);
+    ssize_t   (*write)(void *fs_private, void *file_handle, const void *buf, size_t size);
     int   (*seek)(void *fs_private, void *file_handle, int offset, int whence);
 
     // Directory operations
@@ -115,8 +116,8 @@ bool vfs_umount(const char *mount_path);
 // File operations
 vfs_file_t* vfs_open(const char *path, const char *mode);
 void vfs_close(vfs_file_t *file);
-int vfs_read(vfs_file_t *file, void *buf, int size);
-int vfs_write(vfs_file_t *file, const void *buf, int size);
+int vfs_read(vfs_file_t *file, void *buf, size_t size);
+int vfs_write(vfs_file_t *file, const void *buf, size_t size);
 int vfs_seek(vfs_file_t *file, int offset, int whence);
 int vfs_poll(vfs_file_t *file, struct poll_table *pt);
